requirements.txt
azure-storage-blob==12.19.0
python-dotenv==1.0.0
fastavro==1.9.3
tqdm==4.62.3
pandas==2.1.4
pyarrow==14.0.2
azure-functions==1.18.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.25